% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctlookup-package.R
\docType{package}
\name{ctlookup-package}
\alias{ctlookup}
\alias{ctlookup-package}
\title{ctlookup: A data set of Trial Registration Numbers (TRNs) matched to Digital Object Identifiers (DOIs) for their primary results publications}
\usage{
data(lookup)
}
\description{
Provides a database of Trial Registration Numbers (TRNs) that have been manually matched with Digital Object Identifiers (DOIs).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bgcarlisle/ctlookup}
  \item Report bugs at \url{https://github.com/bgcarlisle/ctlookup/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Gregory Carlisle \email{murph@bgcarlisle.com} (\href{https://orcid.org/0000-0001-8975-0649}{ORCID})

}
\keyword{internal}
